requirements.txt
torch==2.0.1+cpu
transformers==4.36.2
sentence-transformers==2.2.2
huggingface-hub==0.20.3
requests
